Use an SSL certificate with SAN=dns:localhost,ip:127.0.0.1
This gets rid of the extra step where users had to manually accept the
certificate in their browser, which was problematic since it allowed
man-in-the-middle attacks with a certificate that one could just grab from
Github.

With this in place browsers should accept this cert automatically as long as
it's being used from localhost, which generally for funnel is the case.

Also add a script to generate the cert, so that that is documented in case we
need to do it over.